// src/store/enums/StoreEnums.ts
enum Actions {
  // action types
  ADD_BODY_CLASSNAME = "addBodyClassName",
  REMOVE_BODY_CLASSNAME = "removeBodyClassName",
  ADD_BODY_ATTRIBUTE = "addBodyAttribute",
  REMOVE_BODY_ATTRIBUTE = "removeBodyAttribute",
  ADD_CLASSNAME = "addClassName",
  // VERIFY_AUTH = "verifyAuth",
  LOGIN = "login",
  LOGOUT = "logout",
  REGISTER = "register",
  UPDATE_USER = "updateUser",
  FORGOT_PASSWORD = "forgotPassword",
  SET_BREADCRUMB_ACTION = "setBreadcrumbAction",
  Delete_Admin = "deleteAdmin",
  ADD_ADMIN = "addAdmin",
  UPDATE_ADMIN = "updateAdmin",
  ALLADMIN = "All_Admin",
  RESET_CODE = "reset-code",
  CHANGE_PASS="change-pass",
  CHANGE_OLDPASS = "chage-old-pass",
  PROFILE = "profile",
  ALLAREA = "All_area",
  Delete_Area = "delete_area",
  ADD_AREA = "add_area",
  UPDATE_AREA = "update_area",
  EXPORT_AREAS = "export_area",
  IMPORT_AREAS = "IMPORT_AREAS",
  ALLROLES = "all_roles",
  SHOWROLE = "show_role",
  SHOWAREA = "show_areas",
  SHOWADMIN = "show_admin",
  DELETE_ROLE = "delete_role",
  SHOW_ROLE_ADMINS  = "show_role_admins",
  SHOW_PERMISIONS = "show_permisions",
  ADD_ROLE = "add_role",
  UPDATE_ROLE = "update_role",
  ALLCOMPANIES = "all_companies",
  SHOWCOMPANIES = "show_companies",
  Delete_COMPANIES = "delete_companies",
  ADD_COMPANY = "add_company",
  UPDATE_COMPAINES = "update_companies",
  ALLMANAGER = "all_manager",
  ALLBRANCHES = "all_branches",
  ADD_BRANCHES = "add_branches",
  UPDATE_BRANCHES = "update_branch",
  Delete_BRANCHES = "delete_branches",
  SHOWBRANCE = "show_branch",
  ALLCities = "all_cities",
  SHOWMANAGER = "show_manager",
  DELETE_MANAGER = "delete_manager",
  ADD_MANAGER = "add_manager",
  ALLBANKS = "all_banks",
  SHOWBANKS = "show_bankes",
  Delete_BANKES = "delete_bankes",
  ADD_BANK = "add_bank",
  IS_ACTIVE = "is_Active",
  ALLMANAGER_param = "all_manager_doesnt",
  ALLMANAGER_hasActive = "all_manager_has_active",
  UPDATE_BANK = "update_bank",
  Delete_CITIES = "delete_cities",
  SHOWCITIES = "show_cities",
  IMPORT_CITIES = "IMPORT_cities",
  UPDATE_CITIES="update_ciities",
  ADD_CITIES="add_cities",
  ALLServices = "all_services",
  SHOWSERVICES = "show_services",
  Delete_Services = "delete_services",
  ADD_SERVICES = "Add_services",
  UPDATE_SERVICES = "update_services",
  SIGNUP = "SIGN_UP",
  UPDATE_MANAGER = "update_manager",
  ALLPOSUSERS = "all_pos_users",
  SHOWPOSTUSERS = "show_post_users",
  DELETE_POS_USERS = "delete_pos_users",
  Add_ADD_POSUSERS = "add_pos_users",
  UPDATE_POSUSERS = "update_posusers"
  
  
}

enum Mutations {
  // mutation types
  SET_CLASSNAME_BY_POSITION = "appendBreadcrumb",
  PURGE_AUTH = "logOut",
  SET_AUTH = "setAuth",
  SET_USER = "setUser",
  SET_PASSWORD = "setPassword",
  SET_ERROR = "setError",
  SET_BREADCRUMB_MUTATION = "setBreadcrumbMutation",
  SET_LAYOUT_CONFIG = "setLayoutConfig",
  RESET_LAYOUT_CONFIG = "resetLayoutConfig",
  OVERRIDE_LAYOUT_CONFIG = "overrideLayoutConfig",
  OVERRIDE_PAGE_LAYOUT_CONFIG = "overridePageLayoutConfig",
  SET_DELETE  ="DeleteAdmin",
  NEW_ADD_ADMIN = "newAddAdmin",
  SET_UPDATE_ADMIN ="UpdateAdmin",
  SET_ERROR_USER= "seEerrorUser",
  SET_ADMIN="setAdmin",
  SET_TOKEN = "set-token",
  SET_PROFILE = "SET_PROFILE",
  SET_Area = "set_area",
  SET_ERROR_AREA = "Set_error_area",
  SET_DELETE_AREA = "set_delete_area",
  SET_EXPORT_AREAS = "ser_export_areas",
  NEW_ADD_AREA = "new_add",
  SET_UPDATE_Area = "set_update_area",
  SET_ROLES = "set_roles",
  SET_ERROR_ROLES = "set_error_roles",
  SET_SHOW_AREA = "set_show_area",
  SET_SHOW_ROLES = "set_show_roles",
  SET_SHOW_ADMIN = "set_show_admin",
  SET_DELETE_ROLE = "set_delete_role",
  SET_ROLES_ADMINS = "set_Admins_role",
  SET_PERMISIONS = "set_permisions",
  NEW_ADD_ROLE = "new_role",
  SET_UPDATE_ROLE="set_update_role",
  SET_COMPANIES = "set_companies",
  SET_ERROR_COMPANIES = "set_error_companies",
  SET_SHOW_COMPANIES = "set_show_companies",
  SET_DELETE_COMPANIES = "set_delete_companies",
  NEW_ADD_Companies = "new_add_Companies",
  SET_UPDATE_COMPAINES = "set_update_companies",
  SET_BRANCHES = "set_branch",
  SET_ERROR_BRANCHES = "set_error_branches",
  SET_SHOW_BRANCHES = "set_show_branches",
  SET_DELETE_BRANCHES = "set_delete_branches",
  NEW_ADD_BRANCHES = "new_add_branches",
  SET_UPDATE_BRANCHES = "set_update_branches",
  SET_MANAGER = "set_manager",
  SET_ALLCities = "set_all_cities",
  SET_ERROR_Cities= "set_errors_cities",
  SET_ERROR_Manager = "set_error_manager",
  SET_SHOW_MANAGER = "set_show_manager",
  SET_DELETE_MANAGER = "set_delete_manager",
  NEW_ADD_MANAGER = "new_add_manager",
  SET_UPDATE_MANAGER = "set_update_manager",
  SET_Banks = "set_banks",
  SET_SHOW_Banks = "set_show_banks",
  SET_DELETE_Banks = "set_delete_banks",
  NEW_ADD_Banks = "new_add_banks",
  SET_UPDATE_Banks = "set_update_banks",
  SET_ERROR_BANKS = "set_error_bankes",
  SET_DELETE_BANKES = "set_delete_bankes",
  NEW_ADD_BANK = "new_add_bank",
  SET_MANAGER_active = "set_manager_active",
  SET_DELETE_CITIES = "set_delete_cities",
  SET_SHOW_CITIES = "set_show_cities",
  SET_UPDATE_CITIES = "set_update_cities",
  NEW_ADD_CITIES = "new_Add_Cities",
  SET_ALLservices = "set_all_services",
  SET_ERROR_services = "set_error_services",
  SET_SHOW_SERVICES = "set_show_services",
  SET_DELETE_SERVICES = "set_delete_services",
  NEW_ADD_SERVICES = "set_Add_services",
  SET_UPDATE_SERVICES = "set_update_services",
  SET_POSUSERS = "set_posusers",
  SET_ERROR_POSUSERS = "set_error_posusers",
  SET_SHOW_PosUsers = "set_show_posUsers",
  SET_DELETE_POS_USERS = "set_delete_pos_users",
  NEW_ADD_POSUSERS = "new_add_pos_users",
  SET_UPDATE_POSUSERS = "set_update_posusers"

}